Updated Hardcoded username and enhanced user creation
Username: The Model Map username with Session Attributes is replaced with retrieving the username from the Authentication class of the Spring Security. The new todos are created with this username only. Updates performed in WelcomeController and TodoController.

Also, in Spring Security Configuration, the user creation is simplified for adding multiple users. Achieved using refactoring the code with extract methods and inline variables.